+++ libs/cardano-ledger-core/testlib/Test/Cardano/Ledger/Imp/Common.hs
@@ -60,6 +60,7 @@ module Test.Cardano.Ledger.Imp.Common (
   HasStatefulGen (..),
   HasGenEnv (..),
   HasSubState (..),
+  subState,
   R.StatefulGen,
   StateGen (..),
   StateGenM (..),